backend/requirements.txt
# Django y REST framework
Django==5.2.2
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.2,<6.0

# Front
django-cors-headers>=3.15.0

# Soporte GeoDjango / PostGIS
psycopg2-binary>=2.9,<3.0
djangorestframework-gis>=0.18

# Qrs
qrcode==8.2
pillow==11.2.1

requests==2.32.3
